LayoutTests/platform/chromium/http/tests/security/inactive-document-with-empty-security-origin-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL about:blank from frame with URL http://127.0.0.1:8000/security/inactive-document-with-empty-security-origin.html#stop. Domains, protocols and ports must match.

This test passes if it doesn't alert something ugly.
